requirements.txt
aiocontextvars==0.2.2
asn1crypto==0.22.0
botocore==1.11.6
cached-property==1.5.2
conda-pack
contextvars==2.4
dataclasses==0.6
easydict==1.8
einops
ftfy
immutables==0.14
jmespath==0.9.4
loguru==0.5.3
msgpack==1.0.0
msgpack-numpy==0.4.7.1
multimethod==1.4
openai
opencv-contrib-python
Pillow-SIMD==5.3.0.post0
pymemcache
regex
s3transfer==0.1.13
tensorpack
torchpack